refactor: 기존 coin에 속하던 ticker domain 로직 분리
