+++ app/src/main/java/com/coinkiri/coinkiri/data/ticker/mapper/TickerMapper.kt
@@ -2,8 +2,8 @@ package com.coinkiri.coinkiri.data.ticker.mapper
 
 import com.coinkiri.coinkiri.data.ticker.dto.request.TickerRequestDto
 import com.coinkiri.coinkiri.data.ticker.dto.response.TickerResponseDto
-import com.coinkiri.coinkiri.domain.coin.entity.request.TickerRequestEntity
-import com.coinkiri.coinkiri.domain.coin.entity.response.TickerResponseEntity
+import com.coinkiri.coinkiri.domain.ticker.entity.request.TickerRequestEntity
+import com.coinkiri.coinkiri.domain.ticker.entity.response.TickerResponseEntity
 
 fun TickerResponseDto.toTickerEntity(): TickerResponseEntity =
     TickerResponseEntity(
