+++ app/src/main/java/com/coinkiri/coinkiri/data/ticker/repositoryimpl/TickerRepositoryImpl.kt
@@ -3,9 +3,9 @@ package com.coinkiri.coinkiri.data.ticker.repositoryimpl
 import com.coinkiri.coinkiri.data.ticker.datasource.TickerDataSource
 import com.coinkiri.coinkiri.data.ticker.mapper.toTickerEntity
 import com.coinkiri.coinkiri.data.ticker.mapper.toTickerRequestDto
-import com.coinkiri.coinkiri.domain.coin.entity.request.TickerRequestEntity
-import com.coinkiri.coinkiri.domain.coin.entity.response.TickerResponseEntity
-import com.coinkiri.coinkiri.domain.coin.repository.TickerRepository
+import com.coinkiri.coinkiri.domain.ticker.entity.request.TickerRequestEntity
+import com.coinkiri.coinkiri.domain.ticker.entity.response.TickerResponseEntity
+import com.coinkiri.coinkiri.domain.ticker.repository.TickerRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
