+++ app/src/main/java/com/coinkiri/coinkiri/domain/ticker/entity/request/TickerRequestEntity.kt
@@ -1,4 +1,4 @@
-package com.coinkiri.coinkiri.domain.coin.entity.request
+package com.coinkiri.coinkiri.domain.ticker.entity.request
 
 data class TickerRequestEntity(
     val type: String,
