+++ app/src/main/java/com/coinkiri/coinkiri/domain/ticker/repository/TickerRepository.kt
@@ -1,0 +1,10 @@
+package com.coinkiri.coinkiri.domain.ticker.repository
+
+import com.coinkiri.coinkiri.domain.ticker.entity.request.TickerRequestEntity
+import com.coinkiri.coinkiri.domain.ticker.entity.response.TickerResponseEntity
+import kotlinx.coroutines.flow.Flow
+
+interface TickerRepository {
+    fun getTickers(tickerRequestEntity: TickerRequestEntity): Flow<TickerResponseEntity>
+    fun closeConnection()
+}
